Don't check for missing static images
There isn't a good way to do this reliably and it complicates the logic
more than is worth the (buggy) result. Instead, if an image is static we
assume the user knows what they are doing and accept it (although we will
convert to absURL or relURL depending on other parameters passed in).

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>